# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  httperror.ErrorResponse:
    properties:
      details:
        items:
          type: string
        type: array
      message:
        example: corpo inválido
        type: string
      type:
        example: validation
        type: string
    type: object
...
  peca.AlterarQuantidadeReservaPecaRequest:
    properties:
      quantidade:
        example: 3
        type: integer
    required:
    - quantidade
    type: object
...
  peca.ReservaPecaResponse:
    properties:
      id:
        example: 1
        type: integer
      ordem_servico_id:
        example: 10
        type: integer
      peca_id:
        example: 1
        type: integer
      quantidade:
        example: 2
        type: integer
    type: object
  peca.ReservarPecaRequest:
    properties:
      peca_id:
        example: 1
        type: integer
      quantidade:
        example: 2
        type: integer
    required:
    - peca_id
    - quantidade
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API de gestão de ordens de serviço de uma oficina mecânica. Tech Challenge
    Fase 1 - Soat.
  title: Sistema de Oficina Mecânica API
  version: "1.0"
paths:
...
  /v1/ordens-servico/{id}/reservas-pecas:
    post:
      consumes:
      - application/json
      description: Cria ou incrementa a reserva da peça para a OS, respeitando estoque
        mínimo e protegendo concorrência com transação e FOR UPDATE.
      parameters:
      - description: ID da Ordem de Serviço
        in: path
        name: id
        required: true
        type: integer
      - description: Peça e quantidade a reservar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/peca.ReservarPecaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/peca.ReservaPecaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reserva uma peça para uma Ordem de Serviço
      tags:
      - Reservas de Peças
  /v1/ordens-servico/{id}/reservas-pecas/{pecaId}:
    put:
      consumes:
      - application/json
      description: Atualiza a quantidade total da reserva existente, recalculando
        disponibilidade dentro de transação com lock da peça.
      parameters:
      - description: ID da Ordem de Serviço
        in: path
        name: id
        required: true
        type: integer
      - description: ID da peça
        in: path
        name: pecaId
        required: true
        type: integer
      - description: Nova quantidade total reservada
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/peca.AlterarQuantidadeReservaPecaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/peca.ReservaPecaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Altera a quantidade total reservada de uma peça na OS
      tags:
      - Reservas de Peças
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"